Multiple charges changes to models
