+++ src/models/__tests__/offense.test.js
@@ -305,116 +305,22 @@ describe('The offense model', () => {
         .toEqual(expect.arrayContaining(expectedErrors))
     })
 
-    it('the ChargeType field is required', () => {
+    it('the Charges field is required', () => {
       const testData = {
         WasCharged: { value: 'Yes' },
       }
-      const expectedErrors = ['ChargeType.presence.REQUIRED']
+      const expectedErrors = ['Charges.presence.REQUIRED']
 
       expect(validateModel(testData, offense))
         .toEqual(expect.arrayContaining(expectedErrors))
     })
 
-    it('the ChargeType field must have a valid value', () => {
+    it('the Charges field must have at least one item', () => {
       const testData = {
         WasCharged: { value: 'Yes' },
-        ChargeType: { value: 'Test' },
+        Charges: { items: [] },
       }
-      const expectedErrors = ['ChargeType.hasValue.value.inclusion.INCLUSION']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtCharge field is required', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-      }
-      const expectedErrors = ['CourtCharge.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtCharge field must have a value', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-        CourtCharge: 'invalid',
-      }
-      const expectedErrors = ['CourtCharge.hasValue.MISSING_VALUE']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtOutcome field is required', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-      }
-      const expectedErrors = ['CourtOutcome.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtOutcome field must have a value', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-        CourtOutcome: [],
-      }
-      const expectedErrors = ['CourtOutcome.hasValue.MISSING_VALUE']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtDate field is required', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-      }
-      const expectedErrors = ['CourtDate.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtDate field must be a valid date', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-        CourtDate: { year: 3000, month: 13, day: 2 },
-      }
-      const expectedErrors = ['CourtDate.date.date.datetime.INVALID_DATE']
-
-      expect(validateModel(testData, offense))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('CourtDate cannot be before applicant birthdate', () => {
-      const applicantBirthdate = { month: 1, day: 2, year: 1980 }
-      const testData = {
-        WasCharged: { value: 'Yes' },
-        CourtDate: { month: 1, year: 1970, day: 2 },
-      }
-
-      const expectedErrors = [
-        'CourtDate.date.date.datetime.DATE_TOO_EARLY',
-      ]
-
-      expect(validateModel(testData, offense, {
-        applicantBirthdate,
-      }))
-        .toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('CourtDate cannot be in the future', () => {
-      const testData = {
-        WasCharged: { value: 'Yes' },
-        CourtDate: { month: 1, year: 2050, day: 2 },
-      }
-
-      const expectedErrors = [
-        'CourtDate.date.date.datetime.DATE_TOO_LATE',
-      ]
+      const expectedErrors = ['Charges.accordion.MISSING_ITEMS']
 
       expect(validateModel(testData, offense))
         .toEqual(expect.arrayContaining(expectedErrors))
@@ -532,10 +438,18 @@ describe('The offense model', () => {
             zipcode: '10022',
             country: { value: 'United States' },
           },
-          ChargeType: { value: 'Felony' },
-          CourtCharge: { value: 'My charges' },
-          CourtOutcome: { value: 'The outcome' },
-          CourtDate: { year: 2017, day: 3, month: 12 },
+          Charges: {
+            items: [
+              {
+                Item: {
+                  ChargeType: { value: 'Felony' },
+                  CourtCharge: { value: 'My charges' },
+                  CourtOutcome: { value: 'The outcome' },
+                  CourtDate: { year: 2017, day: 3, month: 12 },
+                },
+              },
+            ],
+          },
           WasSentenced: { value: 'Yes' },
           Sentence: {
             Description: { value: 'Something' },
@@ -644,10 +558,18 @@ describe('The offense model', () => {
             zipcode: '10022',
             country: { value: 'United States' },
           },
-          ChargeType: { value: 'Felony' },
-          CourtCharge: { value: 'My charges' },
-          CourtOutcome: { value: 'The outcome' },
-          CourtDate: { year: 2017, day: 3, month: 12 },
+          Charges: {
+            items: [
+              {
+                Item: {
+                  ChargeType: { value: 'Felony' },
+                  CourtCharge: { value: 'My charges' },
+                  CourtOutcome: { value: 'The outcome' },
+                  CourtDate: { year: 2017, day: 3, month: 12 },
+                },
+              },
+            ],
+          },
           WasSentenced: { value: 'No' },
           AwaitingTrial: { value: 'Yes' },
           AwaitingTrialExplanation: { value: 'Something' },
@@ -700,41 +622,11 @@ describe('The offense model', () => {
         .not.toEqual(expect.arrayContaining(expectedErrors))
     })
 
-    it('the ChargeType field is not required', () => {
+    it('the Charges field is not required', () => {
       const testData = {
         WasCharged: { value: 'No' },
       }
-      const expectedErrors = ['ChargeType.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .not.toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtCharge field is not required', () => {
-      const testData = {
-        WasCharged: { value: 'No' },
-      }
-      const expectedErrors = ['CourtCharge.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .not.toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtOutcome field is not required', () => {
-      const testData = {
-        WasCharged: { value: 'No' },
-      }
-      const expectedErrors = ['CourtOutcome.presence.REQUIRED']
-
-      expect(validateModel(testData, offense))
-        .not.toEqual(expect.arrayContaining(expectedErrors))
-    })
-
-    it('the CourtDate field is not required', () => {
-      const testData = {
-        WasCharged: { value: 'No' },
-      }
-      const expectedErrors = ['CourtDate.presence.REQUIRED']
+      const expectedErrors = ['Charges.presence.REQUIRED']
 
       expect(validateModel(testData, offense))
         .not.toEqual(expect.arrayContaining(expectedErrors))
